Improve the logs tool and other fixes
Can take a grep parameter so it can look, for example, only for errors.
Limit its output to 5kb.

Some fixes after refactoring + make older tools use the same pattern.
