+++ apps/dbagent/src/app/api/chat/route.ts
@@ -1,7 +1,6 @@
 import { streamText } from 'ai';
 import { chatSystemPrompt, getModelInstance, getTools } from '~/lib/ai/aidba';
 import { getConnection } from '~/lib/db/connections';
-import { getTargetDbConnection } from '~/lib/targetdb/db';
 
 export const runtime = 'nodejs';
 export const maxDuration = 30;
@@ -31,7 +30,6 @@ export async function POST(req: Request) {
   if (!connection) {
     return new Response('Connection not found', { status: 404 });
   }
-  const targetClient = await getTargetDbConnection(connection.connectionString);
   try {
     const context = chatSystemPrompt;
 
@@ -43,7 +41,7 @@ export async function POST(req: Request) {
       model: modelInstance,
       messages,
       system: context,
-      tools: await getTools(connection, targetClient),
+      tools: await getTools(connection),
       maxSteps: 20,
       toolCallStreaming: true
     });
@@ -51,7 +49,8 @@ export async function POST(req: Request) {
     return result.toDataStreamResponse({
       getErrorMessage: errorHandler
     });
-  } finally {
-    await targetClient.end();
+  } catch (error) {
+    console.error('Error in chat', error);
+    return new Response('Error in chat', { status: 500 });
   }
 }
